Always use the :? parameter substitution parameter with rm
While checks and tests are in place, it is a good pattern to use :? on
a variable when using rm -rf.

The exception being when we are echoing the command, as we would exit
half-way an echo print.

Note, that while set -u should cover for this error, to ensure this
always works, it is simply good practice to always do it, even when not
needed.

Contributes to EMP-416

+++ build.sh
@@ -45,7 +45,7 @@ cleanup()
         exit 1
     fi
 
-    rm -rf "${BUILD_DIR}"
+    rm -rf "${BUILD_DIR:?}"
 }
 
 bootstrap_prepare()
